requirements.txt
# local package
-e .

# external requirements
python==3.9.6
openai==0.27.0
customtkinter==5.1.2
pillow==9.4.0
pyglet==2.0.5